Moved writing validation outside PandasDataset for better comprehension. It is only related to SDMX-ML writing.
