+++ src/pysdmx/io/pd.py
@@ -2,8 +2,6 @@
 
 import pandas as pd
 
-from pysdmx.errors import Invalid
-from pysdmx.model import Schema
 from pysdmx.model.dataset import Dataset
 
 
@@ -22,30 +20,3 @@ class PandasDataset(Dataset, frozen=False, kw_only=True):
     """
 
     data: pd.DataFrame
-
-    def writing_validation(self) -> None:
-        """Structural validation of the dataset."""
-        if not isinstance(self.structure, Schema):
-            raise Invalid(
-                "Dataset Structure is not a Schema. "
-                "Cannot perform operation."
-            )
-        if len(self.data.columns) != len(self.structure.components):
-            raise Invalid("Data columns length must match components length.")
-        columns = set(self.data.columns)
-        components = {comp.id for comp in self.structure.components}
-        if columns != components:
-            difference = columns.symmetric_difference(components)
-            raise Invalid(
-                f"Data columns must match components. "
-                f"Difference: {', '.join(difference)}"
-            )
-
-        if not self.structure.components.dimensions:
-            raise Invalid(
-                "The dataset structure must have at least one dimension."
-            )
-        if not self.structure.components.measures:
-            raise Invalid(
-                "The dataset structure must have at least one measure."
-            )
